project/pyproject.toml
[project]
name = "job-portal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "Django==5.0.7",
  "djangorestframework==3.15.2",
  "psycopg[binary]==3.2.1",
  "python-decouple==3.8",
  "djangorestframework-simplejwt==5.3.1",
]
